Move Pybullet Controllers + IK (#1170)
* Move Pybullet controllers and IK

- To their own respective modules
- Using PyCharm 'move' functionality
