+++ tests/envs/test_pybullet_robots.py
@@ -7,7 +7,9 @@ import pytest
 from predicators.src import utils
 from predicators.src.envs.pybullet_env import create_pybullet_block
 from predicators.src.envs.pybullet_robots import get_kinematic_chain, \
-    pybullet_inverse_kinematics, run_motion_planning
+    run_motion_planning
+from predicators.src.pybullet_helpers.inverse_kinematics import \
+    pybullet_inverse_kinematics
 from predicators.src.pybullet_helpers.robots import FetchPyBulletRobot, \
     create_single_arm_pybullet_robot
 from predicators.src.settings import CFG
